Struct massa_storage::Storage

pub struct Storage { /* private fields */ }

A storage system for objects (blocks, operations…), shared by various components.

Implementations

impl Storage

pub fn create_root() -> Storage

Creates a new Storage instance. Must be called only one time in the execution:

In the main for the node

At the top of the test in tests All others instances of Storage mush cloned from this one suing clone() or clone_without_refs().

pub fn clone_without_refs(&self) -> Self

Clones the object to a new one that has no references

pub fn extend(&mut self, other: Storage)

Efficiently extends the current Storage by consuming the refs of another’s.

pub fn split_off(
&mut self,
blocks: &PreHashSet<BlockId>,
operations: &PreHashSet<OperationId>,
endorsements: &PreHashSet<EndorsementId>
) -> Storage

Efficiently splits off a subset of the reference ownership into a new Storage object. Panics if some of the refs are not owned by the source.

pub fn get_block_refs(&self) -> &PreHashSet<BlockId>

get the block reference ownership

pub fn claim_block_refs(
&mut self,
ids: &PreHashSet<BlockId>
) -> PreHashSet<BlockId>

Claim block references. Returns the set of block refs that were found and claimed.

pub fn drop_block_refs(&mut self, ids: &PreHashSet<BlockId>)

Drop block references

pub fn store_block(&mut self, block: WrappedBlock)

Store a block Note that this also claims a local reference to the block

pub fn claim_operation_refs(
&mut self,
ids: &PreHashSet<OperationId>
) -> PreHashSet<OperationId>

Claim operation references. Returns the set of operation refs that were found and claimed.

pub fn get_op_refs(&self) -> &PreHashSet<OperationId>

get the operation reference ownership

pub fn drop_operation_refs(&mut self, ids: &PreHashSet<OperationId>)

Drop local operation references. Ignores already-absent refs.

pub fn store_operations(&mut self, operations: Vec<WrappedOperation>)

Store operations Claims a local reference to the added operation

pub fn read_operations(&self) -> RwLockReadGuard<'_, OperationIndexes>

Gets a read reference to the operations index

pub fn read_endorsements(&self) -> RwLockReadGuard<'_, EndorsementIndexes>

Gets a read reference to the endorsements index

pub fn read_blocks(&self) -> RwLockReadGuard<'_, BlockIndexes>

Gets a read reference to the blocks index

pub fn claim_endorsement_refs(
&mut self,
ids: &PreHashSet<EndorsementId>
) -> PreHashSet<EndorsementId>

Claim endorsement references. Returns the set of operation refs that were found and claimed.

pub fn get_endorsement_refs(&self) -> &PreHashSet<EndorsementId>

get the endorsement reference ownership

pub fn drop_endorsement_refs(&mut self, ids: &PreHashSet<EndorsementId>)

Drop local endorsement references. Ignores already-absent refs.

pub fn store_endorsements(&mut self, endorsements: Vec<WrappedEndorsement>)

Store endorsements Claims local references to the added endorsements